Handle exception out of main file
