+++ badrequest.py
@@ -1,0 +1,9 @@
+class BadRequest(Exception):
+
+    def __init__(self, message):
+        self.message = message
+
+        super().__init__(message)
+
+    def __str__(self):
+        return self.message
